Convert SettingsActivity into a Fragment
This commit removes the redundant SettingsActivity since a dedicated SettingsFragment is already in place. The hasHR method, previously in SettingsActivity, has been relocated to SettingsSensorsFragment to align with related sensor settings. Activities that declared SettingsActivity as their parent in the AndroidManifest have been temporarily re-parented to MainLayout. A follow-up commit should address this parenting structure more appropriately.
