--- app/src/main/org/runnerup/view/SettingsActivity.java
@@ -1,96 +1,0 @@
-/*
- * Copyright (C) 2012 - 2014 jonas.oreland@gmail.com
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package org.runnerup.view;
-
-import android.content.Context;
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.content.res.Resources;
-import android.os.Bundle;
-import androidx.activity.OnBackPressedCallback;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
-import androidx.preference.Preference;
-import androidx.preference.PreferenceFragmentCompat;
-import androidx.preference.PreferenceManager;
-import org.runnerup.R;
-
-public class SettingsActivity extends AppCompatActivity
-    implements PreferenceFragmentCompat.OnPreferenceStartFragmentCallback {
-
-  public void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    setContentView(R.layout.settings_activity);
-
-    // Ensure that the fragment is added only once
-    if (savedInstanceState == null) {
-      getSupportFragmentManager()
-          .beginTransaction()
-          .setReorderingAllowed(true)
-          .replace(R.id.settings_fragment_container, new SettingsFragment())
-          .commit();
-    }
-
-    Context context = this;
-    getOnBackPressedDispatcher()
-        .addCallback(
-            this,
-            new OnBackPressedCallback(true) {
-              @Override
-              public void handleOnBackPressed() {
-                int count = getSupportFragmentManager().getBackStackEntryCount();
-
-                if (count == 0) {
-                  // if no more fragments to pop, navigate back
-                  Intent intent = new Intent(context, MainLayout.class);
-                  startActivity(intent);
-                  finish();
-                } else {
-                  getSupportFragmentManager().popBackStack();
-                }
-              }
-            });
-  }
-
-  public static boolean hasHR(Context ctx) {
-    Resources res = ctx.getResources();
-    SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(ctx);
-    String btAddress = prefs.getString(res.getString(R.string.pref_bt_address), null);
-    String btProviderName = prefs.getString(res.getString(R.string.pref_bt_provider), null);
-    return btProviderName != null && btAddress != null;
-  }
-
-  @Override
-  public boolean onPreferenceStartFragment(PreferenceFragmentCompat caller, Preference pref) {
-    // Instantiate the new Fragment for the clicked preference screen to show
-    final Bundle args = pref.getExtras();
-    final Fragment fragment =
-        getSupportFragmentManager()
-            .getFragmentFactory()
-            .instantiate(getClassLoader(), pref.getFragment());
-    fragment.setArguments(args);
-    fragment.setTargetFragment(caller, 0);
-    // Replace the existing Fragment with the new Fragment
-    getSupportFragmentManager()
-        .beginTransaction()
-        .replace(R.id.settings_fragment_container, fragment)
-        .addToBackStack(null)
-        .commit();
-    return true;
-  }
-}
